api/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
python-multipart==0.0.24
pydantic==2.12.5
pydantic-settings==2.13.1
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
openai==2.30.0
python-dotenv==1.2.2
Pillow==12.2.0
